Consolidate backend camera implementations
