+++ docs/source/reference.rst
@@ -38,12 +38,10 @@ Cameras
 *******
 
 .. inheritance-diagram::
-   manim.camera.camera
-   manim.camera.moving_camera
-   manim.camera.multi_camera
-   manim.camera.three_d_camera
+   manim.renderer.cairo.camera
+   manim.renderer.opengl.camera
    :parts: 1
-   :top-classes: manim.camera.camera.Camera, manim.mobject.mobject.Mobject
+   :top-classes: manim.renderer.cairo.camera.Camera, manim.mobject.mobject.Mobject, manim.mobject.opengl.opengl_mobject.OpenGLMobject
 
 Mobjects
 ********
