+++ manim/__init__.py
@@ -41,10 +41,6 @@ from .animation.transform import *
 from .animation.transform_matching_parts import *
 from .animation.updaters.mobject_update_utils import *
 from .animation.updaters.update import *
-from .camera.camera import *
-from .camera.moving_camera import *
-from .camera.multi_camera import *
-from .camera.three_d_camera import *
 from .constants import *
 from .manager import *
 from .mobject.frame import *
