+++ manim/manager.py
@@ -16,10 +16,11 @@ if TYPE_CHECKING:
     from ._config.output import OutputSpec
     from ._config.render_session import RenderSessionSpec
     from .animation.animation import Animation
-    from .camera.camera import Camera
     from .mobject.mobject import Mobject, _AnimationBuilder
     from .renderer.cairo import CairoRenderer
-    from .renderer.opengl import OpenGLCamera, OpenGLRenderer
+    from .renderer.cairo.camera import Camera
+    from .renderer.opengl.camera import OpenGLCamera
+    from .renderer.opengl.renderer import OpenGLRenderer
     from .scene.scene import Scene
     from .scene.scene_file_writer import SceneFileWriter
