+++ manim/mobject/types/image_mobject.py
@@ -14,7 +14,6 @@ from PIL.Image import Resampling
 from manim.mobject.geometry.shape_matchers import SurroundingRectangle
 
 from ... import config
-from ...camera.camera import Camera
 from ...constants import *
 from ...mobject.mobject import Mobject
 from ...utils.bezier import interpolate
@@ -34,6 +33,7 @@ if TYPE_CHECKING:
 
     import numpy.typing as npt
 
+    from manim.renderer.cairo.camera import Camera
     from manim.typing import PixelArray, StrPath
