--- manim/renderer/cairo_renderer.py
@@ -1,5 +1,0 @@
-"""Compatibility imports for the Cairo rendering backend."""
-
-from .cairo import CairoRenderer
-
-__all__ = ["CairoRenderer"]
