+++ manim/renderer/opengl/__init__.py
@@ -5,16 +5,24 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from .renderer import OpenGLCamera, OpenGLRenderer
+    from .camera import OpenGLCamera
+    from .renderer import OpenGLRenderer
 
 __all__ = ["OpenGLCamera", "OpenGLRenderer"]
 
 
 def __getattr__(name: str) -> Any:
-    if name not in __all__:
-        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-    from .renderer import OpenGLCamera, OpenGLRenderer
+    value: Any
+    if name == "OpenGLCamera":
+        from .camera import OpenGLCamera
 
-    value = {"OpenGLCamera": OpenGLCamera, "OpenGLRenderer": OpenGLRenderer}[name]
+        value = OpenGLCamera
+    elif name == "OpenGLRenderer":
+        from .renderer import OpenGLRenderer
+
+        value = OpenGLRenderer
+    else:
+        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
     globals()[name] = value
     return value
