--- manim/renderer/opengl_renderer.py
@@ -1,5 +1,0 @@
-"""Compatibility imports for the OpenGL rendering backend."""
-
-from .opengl import OpenGLCamera, OpenGLRenderer
-
-__all__ = ["OpenGLCamera", "OpenGLRenderer"]
