+++ manim/scene/vector_space_scene.py
@@ -10,13 +10,13 @@ from typing import TYPE_CHECKING, Any, cast
 import numpy as np
 
 from manim.animation.creation import DrawBorderThenFill, Group
-from manim.camera.camera import Camera
 from manim.mobject.geometry.arc import Dot
 from manim.mobject.geometry.line import Arrow, Line, Vector
 from manim.mobject.geometry.polygram import Rectangle
 from manim.mobject.graphing.coordinate_systems import Axes, NumberPlane
 from manim.mobject.opengl.opengl_mobject import OpenGLMobject
 from manim.mobject.text.tex_mobject import MathTex, Tex
+from manim.renderer.cairo.camera import Camera
 from manim.utils.config_ops import update_dict_recursively
 
 from .. import config
