+++ manim/utils/hashing.py
@@ -18,9 +18,9 @@ from manim._config import config, logger
 
 if TYPE_CHECKING:
     from manim.animation.animation import Animation
-    from manim.camera.camera import Camera
     from manim.mobject.mobject import Mobject
-    from manim.renderer.opengl import OpenGLCamera
+    from manim.renderer.cairo.camera import Camera
+    from manim.renderer.opengl.camera import OpenGLCamera
     from manim.scene.scene import Scene
 
 __all__ = ["KEYS_TO_FILTER_OUT", "get_hash_from_play_call", "get_json"]
